CHARTS - Fix Visual bugs
