+++ pages/charts.py
@@ -2,20 +2,35 @@ import collections
 import datetime
 import enum
 import warnings
+import math
 
 warnings.simplefilter(action="ignore", category=FutureWarning)
 import operator
 
-import finnhub
 import numpy as np
 import pandas as pd
-import plotly.express as px
 import plotly.graph_objects as go
 import streamlit as st
-import yfinance as yf
-from plotly.subplots import make_subplots
+from charts import finapi
 
-from apikeys import ConfigKey, get_secret
+class NoDividendsException(Exception):
+    pass
+
+eng_fmt = pd.io.formats.format.EngFormatter(accuracy=2, use_eng_prefix=True)
+eng_fmt.ENG_PREFIXES = {
+        0: "",
+        3: "k",
+        6: "M",
+        9: "H",
+        12: "T",
+        15: "Q",
+    }
+
+def custom_round(x):
+    try:
+        return round(x, 3-int(math.floor(math.log10(abs(x)))))
+    except:
+        return x
 
 def render_main():
     st.set_page_config(layout = "wide")
@@ -35,11 +50,7 @@ def render_main():
                     div[data-testid="stSlider"] {
                         display: none;
                     }
-                </style>
-                <button onclick="window.print()">
-                Click me
-                </button>
-                """, unsafe_allow_html=True)
+                </style>""", unsafe_allow_html=True)
     st.title("StockCharts")
     
     ticker_input = st.text_input("Ticker Value")
@@ -57,11 +68,11 @@ def render_main():
     def similar_range_to(df, min1, max1):
         min2, max2 = df.min(), df.max()
     
-        # ranges must be within 2x.
+        # ranges must be within 3x.
         if max2 - min2 != 0:
             ratio = abs((max1 - min1) / (max2 - min2))
     
-            if ratio > 2 or ratio < 0.5:
+            if ratio > 3 or ratio * 3 < 1:
                 return False
     
         # require some intersection
@@ -71,125 +82,28 @@ def render_main():
         # perhaps add more conditions
         return True
         
-    
-    def flatten_report(report):
-        date_index = get_quarter_end_date(report["year"], report["quarter"])
-    
-        return {
-            "index": date_index,
-            **{
-                v["concept"]: v["value"]
-                for sheet in report["report"].values()
-                for v in sheet
-            },
-        }
-    
-    
-    finnhub_client = finnhub.Client(api_key=get_secret(ConfigKey.FINNHUB))
-    SP500DATA = yf.Ticker("^GSPC")
-    
-    def verify_quarterly_data_irregularities(data, start_date):
-        data = data[data.index >= start_date]
-        if data.isna().any():
-            print("DEBUG:")
-            print(data.to_string())
-            st.warning("Requested data contains nans")
-        quarter_ends = pd.date_range(start=start_date, end=data.index[-1], freq='QE').date
-        missing_quarter_ends = [date for date in quarter_ends if date not in data.index]
-        if len(missing_quarter_ends) > 0:
-            print("DEBUG:")
-            print(data.to_string())
-            st.warning(f"Requested data does not contain these monthend values: {missing_quarter_ends}")
-    
     @st.cache_data
     def get_ticker_data(ticker, start_date):
-        yahoo_finance_info = yf.Ticker(ticker)
-        finnhub_data = finnhub_client.company_basic_financials(ticker, "all")
-    
-        data1 = pd.DataFrame(
-            [
-                flatten_report(a)
-                for a in finnhub_client.financials_reported(symbol=ticker, freq="annual")[
-                    "data"
-                ]
-            ]
-        ).set_index("index")[::-1]
-        data2 = pd.DataFrame(
-            [
-                flatten_report(a)
-                for a in finnhub_client.financials_reported(
-                    symbol=ticker, freq="quarterly"
-                )["data"]
-            ]
-        ).set_index("index")[::-1]
-        finnhub_reported_data = pd.concat([data1, data2]).sort_index()
-        finnhub_reported_data.index = pd.to_datetime(finnhub_reported_data.index, utc=True).date
-    
-        hprice = yahoo_finance_info.history(period="max", auto_adjust=False)["Open"]
-        hprice.index = pd.to_datetime(hprice.index).date
-    
-        SP500_hprice = SP500DATA.history(period="max", auto_adjust=False)["Open"]
-        SP500_hprice.index = pd.to_datetime(SP500_hprice.index).date
-    
-        mapping = collections.defaultdict(dict)
-        for k, v in finnhub_data["series"]["quarterly"].items():
-            for i in v:
-                mapping[i["period"]][k] = i["v"]
-    
-        quarterly_series = pd.DataFrame(mapping).transpose()[::-1]
-        quarterly_series.index = np.vectorize(get_quarter_end_by_date)(pd.to_datetime(quarterly_series.index).date)
-    
-        price_to_book = quarterly_series["pb"]
-        book_value = quarterly_series["bookValue"]
-    
-        q_mcap = (book_value.mul(price_to_book, fill_value=np.NaN) * (10**6))
-        q_mcap.index = pd.to_datetime(q_mcap.index).date
-    
-        # so this may look stupid, multiply and divide by the same column, but this essentially is first dividing by
-        # the values on the exact dates to 'normalize' the value, and then interpolate with the remaining prices as indexes.
-        mcap = piecewise_op_search(
-            hprice, piecewise_op_search(q_mcap, hprice, operator.truediv), operator.mul
-        )
-    
-        num_shares = nan_default_chain(
-            finnhub_reported_data,
-            [
-                "us-gaap_WeightedAverageNumberOfSharesOutstandingBasic",
-                "WeightedAverageNumberOfSharesOutstandingBasic",
-            ],
-        )
-    
-        dividend_amount = yahoo_finance_info.dividends
-        dividend_amount.index = pd.to_datetime(dividend_amount.index).date
-    
-        # sum dividends by quarter
-    
-    
-        revenue = nan_default_chain(
-                finnhub_reported_data,
-                ["us-gaap_Revenues", "Revenues", "us-gaap_RevenueFromContractWithCustomerExcludingAssessedTax", "RevenueFromContractWithCustomerExcludingAssessedTax"],
-                fill_zeros=True,
-                adjust=AdjustReported.ALLYTD
-        )
-    
-        q_ev = quarterly_series["ev"]
-        q_ev = q_ev[q_ev.index >= start_date] * 1000000
-    
+        hprice = finapi.price(ticker)
+        SP500_hprice = finapi.price("^GSPC")
+        mcap = finapi.mcap(ticker)
+        dividend_amount = finapi.dividends(ticker)
+        revenue = finapi.revenue(ticker)
+        q_ev = finapi.enterprise_value(ticker)
+
+        # TODO: we should adjust to start of quarter
         ev = piecewise_op_search(q_ev, mcap, operator.sub)
         ev = piecewise_op_search(mcap, ev, operator.add)
     
         return {
             "hprice": hprice,
-            "num_shares": num_shares,
-            "ticker_name": ticker,
+            "ticker": ticker,
             "start_date": start_date,
             "ev": ev,
             "dividend_amount": dividend_amount,
             "revenue": revenue,
             "SP500_hprice": SP500_hprice,
-            "quarterly_series": quarterly_series,
             "mcap": mcap,
-            "finnhub_reported_data": finnhub_reported_data,
         }
     
     
@@ -210,40 +124,10 @@ def render_main():
         temp_df.index = df1.index
         return op(df1, temp_df)
     
-    
-    class AdjustReported(enum.Enum):
-        NOTHING, ANNUALSYTD, ALLYTD = range(3)
-    
-    
-    def nan_default_chain(data, key_list, fill_zeros=False, adjust=AdjustReported.NOTHING):
-        key_list = [key for key in key_list if key in data]
-        ret = pd.Series(np.nan, index=data.index)
-        for key in key_list:
-            ret = ret.fillna(data[key])
-        match adjust:
-            # todo: add interpolation
-            case AdjustReported.ANNUALSYTD:
-                ret = ret.rolling(4).apply(
-                    lambda x: x[3] if x.index[3].month != 12 else x[3] - x[0] - x[1] - x[2]
-                )
-            case AdjustReported.ALLYTD:
-                ret = ret.rolling(2).apply(
-                    lambda x: x[1] - x[0] if x.index[1].month != 3 else x[1]
-                )
-            case AdjustReported.NOTHING:
-                pass
-        if fill_zeros:
-            ret = ret.fillna(0)
-        return ret
-    
-    
-    # todo: multiple axes with https://stackoverflow.com/questions/65037641/plotly-how-to-add-multiple-y-axes
-    # todo: add 'current' boxes with https://plotly.com/python/indicator/
-    # todo: clear up naming and [::-1]s, make functions less leaky
-    
     # graph1
     @st.cache_data
-    def get_graph1(*_, start_date, hprice, ticker_name, SP500_hprice, **rest):
+    def get_graph1(*_, start_date, hprice, ticker, SP500_hprice, **rest):
+        print(hprice.index.searchsorted(start_date))
         price_return = (
             ((hprice / hprice[hprice.index.searchsorted(start_date)]) - 1)
         )
@@ -251,7 +135,7 @@ def render_main():
             ((SP500_hprice / SP500_hprice[SP500_hprice.index.searchsorted(start_date)]) - 1)
         )
     
-        graph1 = pd.concat([SP500_return, price_return], keys=["SP 500", ticker_name], axis=1)
+        graph1 = pd.concat([SP500_return, price_return], keys=["SP 500", ticker], axis=1)
         graph1.index = pd.to_datetime(graph1.index, utc=True).date
     
         return graph1
@@ -259,7 +143,7 @@ def render_main():
     
     # graph2
     @st.cache_data
-    def get_graph2(*_, quarterly_series, mcap, ev, **rest):
+    def get_graph2(*_, mcap, ev, **rest):
         graph2 = pd.concat([ev, mcap], keys=["Enterprise Value", "Market Cap"], axis=1)
         graph2.index = pd.to_datetime(graph2.index, utc=True).date
     
@@ -269,41 +153,21 @@ def render_main():
     # graph3
     # For IBM, 10-k is missing in finnhub API, causing issues. However given finnhub is right this should be right
     @st.cache_data
-    def get_graph3(*_, finnhub_reported_data, start_date, **rest):
-        eps_diluted = nan_default_chain(
-            finnhub_reported_data,
-            [
-                "us-gaap_EarningsPerShareDiluted",
-                "EarningsPerShareDiluted",
-                "us-gaap_EarningsPerShareBasic",
-                "EarningsPerShareBasic",
-            ],
-            fill_zeros=True,
-            adjust=AdjustReported.ALLYTD
-        )
-        raw = nan_default_chain(
-            finnhub_reported_data,
-            [
-                "us-gaap_EarningsPerShareDiluted",
-                "EarningsPerShareDiluted",
-                "us-gaap_EarningsPerShareBasic",
-                "EarningsPerShareBasic",
-            ],
-            fill_zeros=True,
-            adjust=AdjustReported.NOTHING
-        )
-        print("HELLO")
-        print(raw)
-        verify_quarterly_data_irregularities(eps_diluted, start_date.replace(year = start_date.year - 1))
-        graph3 = eps_diluted.rolling(4).sum().rename("EPS Diluted (TTM)").to_frame()
+    def get_graph3(*_, ticker, start_date, **rest):
+        eps_diluted = finapi.eps_diluted(ticker)
+        # TODO: make sure rollings are expected direction
+        graph3 = eps_diluted.rolling(4).sum().rename("EPS Diluted Before Extra (TTM)").to_frame()
     
         return graph3
     
     
     # graph4
     @st.cache_data
-    def get_graph4(*_, finnhub_reported_data, num_shares, dividend_amount, quarterly_series, start_date, **rest):
-        dividend_ratios = dividend_amount.rolling(4).sum().rolling(5).apply(lambda a: (a[4] - a[0]) / a[0] * 100)
+    def get_graph4(*_, dividend_amount, start_date, **rest):
+        if dividend_amount is None:
+            raise NoDividendsException("No dividends found")
+
+        dividend_ratios = dividend_amount.rolling(4).sum().rolling(5).apply(lambda a: (a[4] - a[0]) / a[0])
     
         graph4 = pd.concat(
             [dividend_ratios, dividend_amount],
@@ -315,16 +179,11 @@ def render_main():
     
     
     # graph5
-    # good
     @st.cache_data
-    def get_graph5(*_, quarterly_series, mcap, start_date, **rest):
-        price_to_book = quarterly_series["pb"]
-        price_to_cashflow = quarterly_series["pfcfTTM"]
-        price_to_sales = quarterly_series["psTTM"]
-    
-        verify_quarterly_data_irregularities(price_to_book, start_date)
-        verify_quarterly_data_irregularities(price_to_cashflow, start_date)
-        verify_quarterly_data_irregularities(price_to_sales, start_date)
+    def get_graph5(*_, ticker, mcap, start_date, **rest):
+        price_to_book = finapi.price_book(ticker) 
+        price_to_cashflow = finapi.price_cashflow(ticker).rolling(4).mean() / 4
+        price_to_sales = finapi.price_sales(ticker).rolling(4).mean() / 4
     
         graph5 = pd.concat([price_to_book, price_to_cashflow, price_to_sales], keys=["Price to Book", "Price to Cashflow (TTM)", "Price to Sales (TTM)"], axis=1)
         graph5.index = pd.to_datetime(graph5.index, utc=True).date
@@ -339,27 +198,28 @@ def render_main():
     # graph6
     # ebit issue is issue with IBM
     @st.cache_data
-    def get_graph6(*_, finnhub_reported_data, quarterly_series, ev, start_date, num_shares, **rest):
-        verify_quarterly_data_irregularities(quarterly_series["ebitPerShare"], start_date)
-        ebit = piecewise_op_search(
-                quarterly_series["ebitPerShare"],
-            num_shares,
-            operator.mul,
-            )
-    
-        ebit = ebit.rolling(4).sum()
-    
-        ev_ebit = piecewise_op_search(ev, ebit, operator.truediv)
-        graph6 = ev_ebit.rename("EV / EBIT (TTM)").to_frame()
+    def get_graph6(*_, ticker, ev, **rest):
+        ebitda = finapi.ebitda(ticker).rolling(4).sum()
+        ev_ebitda = piecewise_op_search(ev, ebitda, lambda df1, df2: df1.div(df2, axis=0))
+
+        ebit_revenue = finapi.ebit_revenue(ticker)
+        revenue = finapi.revenue(ticker)
+        ebit = piecewise_op_search(ebit_revenue, revenue, lambda df1, df2: df1.mul(df2, axis=0)).rolling(4).sum()
+
+        ev_ebit = piecewise_op_search(ev, ebit, lambda df1, df2: df1.div(df2, axis=0))
+        graph6 = pd.concat([ev_ebit, ev_ebitda], keys=["EV / EBIT (TTM)", "EV / EBITDA (TTM)"], axis=1)
     
         return graph6
     
     
     # graph7
     @st.cache_data
-    def get_graph7(*_, finnhub_reported_data, hprice, ticker_name, dividend_amount, start_date, num_shares, **rest):
+    def get_graph7(*_, hprice, ticker, dividend_amount, start_date, **rest):
+        if dividend_amount is None:
+            raise NoDividendsException("No dividends found")
+
         # todo: move into display logic?
-        price_return = hprice.rename(ticker_name)[hprice.index >= start_date]
+        price_return = hprice.rename(ticker)[hprice.index >= start_date]
     
         dividend_amount = dividend_amount[dividend_amount.index >= start_date]
     
@@ -377,29 +237,33 @@ def render_main():
     
     # graph8
     @st.cache_data
-    def get_graph8(*_, finnhub_reported_data, quarterly_series, revenue, start_date, num_shares, **rest):
+    def get_graph8(*_, ticker, revenue, start_date, **rest):
         revenue_ttm = revenue.rolling(4).sum()
-        revenue_growth = revenue_ttm.rolling(5).apply(lambda a: a[4] - a[0])
-        revenue_growth = (revenue_growth / revenue_growth[revenue_growth.index.searchsorted(start_date)]) - 1
+        revenue_growth = revenue_ttm.rolling(5).apply(lambda a: (a[4] / a[0]) - 1)
+        
+        ebitda_ttm = finapi.ebitda(ticker).rolling(4).sum()
     
-        graph8 = revenue_growth.rename("Revenue Growth (YoY)").to_frame()
+        graph8 = pd.concat([revenue_growth, ebitda_ttm], keys=["Revenue Growth (YoY)", "EBITDA (TTM)"], axis=1)
         return graph8
     
     
     # graph9
     @st.cache_data
-    def get_graph9(*_, quarterly_series, **rest):
+    def get_graph9(*_, ticker, **rest):
+        roe = finapi.return_equity(ticker).rolling(4).sum()
+        rotc = finapi.return_total_capital(ticker).rolling(4).sum()
+        roa = finapi.return_assets(ticker).rolling(4).sum()
     
         graph9 = pd.concat(
             [
-                quarterly_series["roeTTM"],
-                quarterly_series["rotcTTM"],
-                quarterly_series["roaTTM"],
+                roe,
+                rotc,
+                roa
             ],
             keys=[
-                "RO Equity",
-                "RO Total Capital",
-                "RO Assets",
+                "Return on Equity (TTM)",
+                "Return on Total Capital (TTM)",
+                "Return on Assets (TTM)",
             ],
             axis=1,
         )
@@ -409,25 +273,18 @@ def render_main():
     
     # graph10
     @st.cache_data
-    def get_graph10(*_, finnhub_reported_data, revenue, quarterly_series, **rest):
+    def get_graph10(*_, ticker, revenue, **rest):
         revenue_ttm = revenue.rolling(4).sum()
+        net_income = finapi.net_income(ticker).rolling(4).sum()
     
-        graph10 = pd.concat([revenue_ttm], keys=['Revenue (TTM)'], axis=1)
+        graph10 = pd.concat([revenue_ttm, net_income], keys=['Total Revenue (TTM)', 'Net Income (TTM)'], axis=1)
         return graph10
     
     
     # graph11
     @st.cache_data
-    def get_graph11(*_, finnhub_reported_data, **rest):
-        shares_outstanding_diluted = nan_default_chain(
-            finnhub_reported_data,
-            [
-                "us-gaap_WeightedAverageNumberOfDilutedSharesOutstanding",
-                "WeightedAverageNumberOfDilutedSharesOutstanding",
-            ],
-        )
-    
-        graph11 = shares_outstanding_diluted.rolling(4).mean().rename("Weighted Average Diluted Shares Outstanding (TTM)").to_frame()
+    def get_graph11(*_, ticker, **rest):
+        graph11 = finapi.weighted_average_shares_diluted_outstanding(ticker).rolling(4).mean().rename("Weighted Average Diluted Shares Outstanding (TTM)").to_frame()
     
         return graph11
     
@@ -435,12 +292,13 @@ def render_main():
     # graph12
     # sac tangible, book value and book value per share
     @st.cache_data
-    def get_graph12(*_, finnhub_reported_data, quarterly_series, num_shares, **rest):
-        book_value = quarterly_series["bookValue"] * 1000000
+    def get_graph12(*_, **rest):
+        book_share = finapi.book_value_share(ticker)
+        tangible_bv = finapi.tangible_book_value(ticker)
+        total_debt = finapi.total_debt(ticker)
     
-        book_share = piecewise_op_search(book_value, num_shares, operator.truediv)
-    
-        graph12 = pd.concat([book_share, book_value], keys=["BV per share", "Book Value"], axis=1)
+        graph12 = pd.concat([book_share, tangible_bv, total_debt], keys=["BV per share", "Tangible Book Value", "Total Debt"], axis=1)
+
     
         return graph12
     
@@ -468,43 +326,45 @@ def render_main():
             "singleaxis": True,
         },
         {
-            "title": "Market Cap, Enterprise Value",
+            "shortform": { "Enterprise Value": "EV"}
         },
         {
-            "title": "EPS: Diluted Before Extra (TTM)",
+            "shortform": { "EPS Diluted Before Extra (TTM)": "Diluted EPS" },
             "show_legend": False,
         },
         {
-            "title": "1 Year Dividend Growth Rate, Dividend Amount",
+            "shortform": { "1 Year Dividend Growth Rate": "Div Growth", "Dividend Amount": "Div Amount"},
+            "ypercent": [True, False],
         },
         {
-            "title": "Price / Book, Price / Cash Flow (TTM), Price / Sales (TTM)",
+            "shortform": { "Price to Book": "P/Book", "Price to Cashflow (TTM)": "P/Cashflow", "Price to Sales (TTM)": "P/Sales" },
         },
         {
-            "title": "EV / EBIT (TTM)",
+            "shortform": { "EV / EBIT (TTM)": "EV/EBIT", "EV / EBITDA (TTM)": "EV/EBITDA" }
         },
         {
-            "title": "Price Return, Total Return",
             "ypercent": True,
             "singleaxis": True,
         },
         {
-            "title": "Revenue Growth (YoY)",
-            "ypercent": True,
+            "shortform": { "Revenue Growth (YoY)": "Revenue Growth", "EBITDA (TTM)": "EBITDA" },
+            "ypercent": [True, False],
+            "singleaxis": False,
         },
         {
-            "title": "Return on Equity (TTM), Return on Total Capital (TTM), Return on Assets (TTM)",
-            "ypercent": True,
+            "shortform": { "Return on Equity (TTM)": "RoE", "Return on Total Capital (TTM)": "RoTC", "Return on Assets (TTM)": "RoA" },
+            "ypercent": True
         },
         {
-            "title": "Total Revenue (TTM)",
+            "shortform": { "Total Revenue (TTM)": "Revenue", "Net Income (TTM)": "Income" }
         },
         {
-            "title": "Diluted Weighted Average Shares Outstanding (TTM)",
+            "shortform": { "Diluted Weighted Average Shares Outstanding (TTM)": "DWAS Outstanding" },
             "show_legend": False,
         },
         {
-            "title": "Book Value / Share, Book Value",
+            "title": "Book Value / Share, Tangible Book Value, Total Debt",
+            "shortform": { "Book Value / Share": "BV/Share", "Tangible Book Value": "Tang BV", "Total Debt": "Debt"}
         },
     ]
     
@@ -549,7 +409,7 @@ def render_main():
     for func, data in zip(graph_funcs, meta_info):
        try:
            graph = func(**ticker_data)
-           graph = graph[graph.index >= start_date]
+           graph = graph[graph.index >= start_date].applymap(custom_round)
     
            col1, col2 = st.columns(2)
            col2.text("\n")
@@ -558,13 +418,18 @@ def render_main():
            col2.text("\n")
            col2.text("\n")
            col2.text("\n")
-           col2.dataframe(graph.iloc[-1])
+           col2.dataframe(graph.iloc[[0, -1]].style.format({
+               col: '{:,.2%}'.format if "ypercent" in data and (data["ypercent"] is True or data["ypercent"][i] is True) else eng_fmt
+               for i, col in enumerate(graph)
+               }))
            
            fig = go.Figure()
            layoutupdate = {}
            axes = []
            for col in graph:
                for i, axis in enumerate(axes):
+                   if ("singleaxis" in data and not data["singleaxis"]):
+                       continue
                    if ("singleaxis" in data and data["singleaxis"]) or similar_range_to(graph[col], axis[1], axis[2]):
                        axes[i] = ([*axes[i][0], col], min(axis[1], graph[col].min()), max(axis[2], graph[col].max()))
                        break
@@ -574,23 +439,27 @@ def render_main():
            for i, axis in enumerate(axes):
                label = str(i + 1) if i != 0 else ""
                for col in axis[0]:
-                   fig.add_trace(go.Scatter(x=graph.index, y=graph[col], name=col,yaxis=f"y{label}"))
-               layoutupdate[f"yaxis{label}"] = dict(title=', '.join(axis[0]), autoshift=True, title_standoff=10, shift=-30)
+                   print((col if not ('shortform' in data and col in data['shortform']) else data['shortform'][col]))
+                   fig.add_trace(go.Scatter(x=graph.index, y=graph[col],hovertemplate='%{y}', name=(col if not ('shortform' in data and col in data['shortform']) else data['shortform'][col]),yaxis=f"y{label}"))
+               axis_title = ", ".join([data['shortform'][col] if ('shortform' in data and col in data['shortform']) else col for col in axis[0]])
+               layoutupdate[f"yaxis{label}"] = dict(title=(', '.join(axis[0]) if len(axis[0]) == 1 else axis_title), autoshift=True, title_standoff=10, shift=-30)
                if i != 0:
                    layoutupdate[f"yaxis{label}"]["tickmode"] = "sync"
                    layoutupdate[f"yaxis{label}"]["overlaying"] = "y"
                    layoutupdate[f"yaxis{label}"]["anchor"] = "free"
     
-               if "ypercent" in data:
+               if "ypercent" in data and (data["ypercent"] is True or data["ypercent"][i] is True):
                    layoutupdate[f"yaxis{label}"]["tickformat"] =".2%"
     
-           fig.update_layout(**layoutupdate, title=f"{ticker}: {data['title']}", hovermode="x", xaxis_title="Date", margin_l = 80 + 20 * len(axes))
+           fig.update_layout(**layoutupdate, title=f"{ticker}: {', '.join(graph.columns) if not ('title' in data) else data['title']}", hovermode="x unified", xaxis_title="Date", hoverlabel_namelength=4, margin_l = 80 + 20 * len(axes))
     
            if "show_legend" in data:
                fig.update_layout(showlegend=data["show_legend"])
                fig.update_layout({"legend_title_text": "Legend"})
            col1.plotly_chart(fig, use_container_width=True)
-    
+       except NoDividendsException as nde:
+           st.warning(f"Graph could not be shown as no dividends were found for ticker {ticker}")
+
        except Exception as e:
            st.text(f"Graph {data['title']} could not be shown due to the following error:")
            st.exception(e)
